add topic trie

+++ pkg/packets/packets.go
@@ -55,9 +55,9 @@ const (
 
 // QoS levels & Subscribe failure
 const (
-	QOS_0             = 0x00
-	QOS_1             = 0x01
-	QOS_2             = 0x02
+	QOS_0             = uint8(0x00)
+	QOS_1             = uint8(0x01)
+	QOS_2             = uint8(0x02)
 	SUBSCRIBE_FAILURE = 0x80
 )
